Move syllable hashing utilities to classifiers
